requirements.txt
# Python >= 3.11

resource
websockets
websocket-client
